requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-hashid-field==3.4.1
django-hashids==0.7.1
djangorestframework==3.17.1
drf-yasg==1.21.15
gunicorn==25.3.0
hashids==1.3.1
inflection==0.5.1
mysqlclient==2.2.8
packaging==26.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.12.0
